@@ -6,16 +6,21 @@
   "private": true,
   "type": "module",
   "dependencies": {
-    "@tuleap/tlp-fetch": "workspace:*",
-    "@tuleap/vue2-gettext-init": "workspace:*",
-    "vue": "^2.7.8"
+    "@tuleap/fault": "workspace:*",
+    "@tuleap/fetch-result": "workspace:*",
+    "@tuleap/vue3-gettext-init": "workspace:*",
+    "vue": "3.3.4",
+    "vue3-gettext": "^2.4.0"
   },
   "devDependencies": {
     "@tuleap/build-system-configurator": "workspace:*",
-    "@tuleap/po-gettext-plugin": "workspace:*"
+    "@tuleap/po-gettext-plugin": "workspace:*",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "neverthrow": "6.0.0"
   },
   "scripts": {
-    "build": "webpack --config webpack.prod.js",
-    "watch": "webpack --config webpack.dev.js --watch"
+    "typecheck": "vue-tsc --noEmit",
+    "build": "vite build",
+    "watch": "vite build --watch --mode development --minify false"
   }
 }
