@@ -201,10 +201,12 @@ class SvnRouter implements DispatchableWithRequest
                     if (! $request->getCurrentUser()->isAdmin($request->getProject()->getID())) {
                         $GLOBALS['Response']->send400JSONErrors(
                             [
-                                'error' => dgettext(
-                                    'tuleap-svn',
-                                    "You don't have permissions to see user groups."
-                                ),
+                                'error' => [
+                                    'message' => dgettext(
+                                        'tuleap-svn',
+                                        "You don't have permissions to see user groups."
+                                    ),
+                                ],
                             ]
                         );
                     }
